logs actual error loading bootstrap server (#14)
If communication fails between the connection manager and the hub cluster, the
resulting log statement only says "failed to load bootstrap server". It does
not include any information about why. This change adds the contents of the
error object to the log statement.

I ran into this problem myself, where communication was failing due to having
an insufficient kubeconfig. It was hard to know what was going on until I added
this information to the log output.