Handle updates to clusterDeployment name and namespace
When clusterDeployment name is updated, the reference in syncset is updated
and propagation is handled by syncset. Since a syncset is always created in
the namespace of the clusterDeployment, anytime namespace is changed,
syncset needs to be deleted from old namespace and recreated in the new
namespace if it's specified.