CR to access submariner config in managedcluster-broker ns
In order to support Globalnet functionality with Submariner addon
some configuration (aka brokers.submariner.io object) needs to be
stored in the managedclusterset-broker namespace. This configuration
is clusterSet wide and will be used by submariner-addon code when new
ManagedClusters are added to the ManagedClusterSet.

Currently, the clusterRole assigned to the ManagedClusterSet administrator
(i.e., open-cluster-management:managedclusterset:admin:<clusterSetName>)
does not allow creation of the brokers.submariner.io object in the
broker namespace. This PR includes the corresponding role.

Related to: https://github.com/stolostron/backlog/issues/19293

Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>